% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab7task7.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire data}
\format{
## fire
A data frame with 15 rows and 3 columns:
\describe{
  \item{Distance}{Damage}

}
}
\usage{
fire
}
\description{
fire data
}
\keyword{datasets}
