% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterations}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a bar plot of a binomial experiment
}
\description{
Binomial simulation
}
\examples{
mybin(iter=100,n=10, p=0.7)
}
