% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample}

\item{fun}{population}

\item{alpha}{percentage of bootstrap interval}

\item{cx}{character size}

\item{...}{any number of arguments}
}
\value{
bootstrap function to make a number of plots, point
and interval estimates and useful output in the form of a list.
}
\description{
Title
}
\examples{
samA=rnorm(25,mean=25,sd=10)
myboot2(iter=10000,alpha=0.05,samA, cx=1.5,xlab="mean")
}
