% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Title}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{a integer}

\item{iter}{a integer}

\item{lambda}{a integer}

\item{...}{}
}
\value{
graph of CLT Possion
}
\description{
Title
}
\examples{
mycltp(n=10,iter=10000)
}
