% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Mycurve function}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{a numeric vector.}

\item{mu}{a numeric vector.}

\item{sigma}{a numeric vector.}
}
\value{
plot
}
\description{
Mycurve function
}
\examples{
myncurve(5,10,3)
}
