% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquad.R
\name{myquad}
\alias{myquad}
\title{A quadratic function}
\usage{
myquad(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
A quadratic function
}
\examples{
myquad(x = 1:10)
}
