% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the number of seats in the flight}

\item{gamma}{the probability a plane will be truly overbooked}

\item{p}{the probability of a "show"}
}
\value{
a named list containing nd, nc, N, p and gamma -
where nd is calculated using the discrete distribution
and nc is the same calculated with normal approximation.
}
\description{
Title
}
\examples{
ntickets(400, 0.02, 0.95)
}
